Calculate Descriptive Statistics

This function calculates common descriptive statistics for a numeric vector, including mean, median, standard deviation, quartiles, and more.

Usage

descriptive_stats(x, na.rm = TRUE)

Arguments

x

A numeric vector

na.rm

Logical. Should missing values be removed? Default is TRUE.

Value

A list containing descriptive statistics:

mean

The arithmetic mean

median

The median value

sd

Standard deviation

var

Variance

min

Minimum value

max

Maximum value

q25

First quartile (25th percentile)

q75

Third quartile (75th percentile)

n

Number of observations

missing

Number of missing values

Examples

# Generate some sample data
data <- rnorm(100, mean = 50, sd = 10)

# Calculate descriptive statistics
stats <- descriptive_stats(data)
print(stats)
#> $mean
#> [1] 51.30597
#> 
#> $median
#> [1] 51.82166
#> 
#> $sd
#> [1] 9.933397
#> 
#> $var
#> [1] 98.67238
#> 
#> $min
#> [1] 23.87666
#> 
#> $max
#> [1] 76.82557
#> 
#> $q25
#>      25% 
#> 45.36584 
#> 
#> $q75
#>     75% 
#> 58.5016 
#> 
#> $n
#> [1] 100
#> 
#> $missing
#> [1] 0
#> 
#> attr(,"class")
#> [1] "descriptive_stats"